+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 _Copyright (c) 2013 Daniele Veneroni._  
 _Released under GPLv3 License. See [LICENSE.md](LICENSE.md) for further information._
 
+### Update 30/11/2013 (2)
+
+* Add filter frequency range control
+* Changed spectrum color settings using 2 HTML5 color inputs
+* Changed Factory Preset order
+* Configure controls on startup
+
 ### Update 30/11/2013
 
 * Fixed resizing on window resize